refactor: tests to dedicated homepage filter station test file
